Внёс поправки
