+++ src/components/app/app.jsx
@@ -2,39 +2,43 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {BrowserRouter, Route, Switch} from 'react-router-dom';
 
-import Main from '../main/main';
-import MainEmpty from '../main/main-empty/main-empty';
-import PlaceProperty from '../place-property/place-property';
-import Login from '../login/login';
-import Favorites from '../favorites/favorites';
-import FavoritesEmpty from '../favorites/favorites-empty/favorites-empty';
-import Page404 from '../page-404/page-404';
+import Main from '../../pages/main/main';
+import MainEmpty from '../../pages/main/empty/empty';
+import PlaceProperty from '../../pages/place-property/place-property';
+import Login from '../../pages/login/login';
+import Favorites from '../../pages/favorites/favorites';
+import FavoritesEmpty from '../../pages/favorites/empty/empty';
+import Page404 from '../../pages/page-404/page-404';
+
+import {Routes} from './../../const';
+
+const {MAIN: pathMain, MAIN_EMPTY: pathMainEmpty, OFFER: pathOffer, LOGIN: pathLogin, FAVOR: pathFavor, FAVOR_EMPTY: pathFavorEmpty} = Routes;
 
 const App = ({offersCount, places, auth, userName}) => (
   <BrowserRouter>
     <Switch>
 
-      <Route exact path="/">
+      <Route exact path={pathMain}>
         <Main offersCount={offersCount} auth={auth} userName={userName} places={places} />
       </Route>
 
-      <Route exact path="/dev-main-empty">
+      <Route exact path={pathMainEmpty}>
         <MainEmpty auth={auth} userName={userName} />
       </Route>
 
-      <Route exact path="/offer:id?">
+      <Route exact path={pathOffer}>
         <PlaceProperty auth={auth} userName={userName} />
       </Route>
 
-      <Route exact path="/login">
+      <Route exact path={pathLogin}>
         <Login />
       </Route>
 
-      <Route exact path="/favorites">
+      <Route exact path={pathFavor}>
         <Favorites auth={auth} userName={userName} />
       </Route>
 
-      <Route exact path="/dev-favorites-empty">
+      <Route exact path={pathFavorEmpty}>
         <FavoritesEmpty auth={auth} userName={userName} />
       </Route>
