+++ src/const.js
@@ -1,0 +1,9 @@
+export const Routes = {
+  MAIN: `/`,
+  MAIN_EMPTY: `/dev-main-empty`,
+  OFFER: `/offer:id?`,
+  LOGIN: `/login`,
+  FAVOR: `/favorites`,
+  FAVOR_EMPTY: `/dev-favorites-empty`
+}
+;
